Updating and adding sidescroll example

+++ interpolators.go
@@ -12,7 +12,7 @@ func interpolatorUpdater(pri int) *listTask {
 	return &listTask{BasicTask: NewBasicTask(pri), f: func(l ListItem) {
 		it := l.(Interpolator)
 		if !it.IsFrozen() {
-			it.Update(time.Since(timerUpdate.t).Seconds() * 1000.0)
+			it.Update(float32(time.Since(timerUpdate.t).Seconds() * 1000.0))
 		}
 	}}
 }
@@ -22,44 +22,44 @@ func interpolatorUpdater(pri int) *listTask {
 // not be updated, they could also be killed and then removed.
 //
 // IsFrozen() should handle whether or not the Interpolator is frozen
-// Update(dT float64) will be called with dT being the duration since the last
+// Update(dT float32) will be called with dT being the duration since the last
 //                    call to update
 // Kill() being called should result in IsAlive() returning false otherwise IsAlive()
 //        should return true
 type Interpolator interface {
 	IsFrozen() bool
-	Update(dT float64)
+	Update(dT float32)
 	Kill()
 	IsAlive() bool
-	GetValue() float64
+	GetValue() float32
 }
 
 type timeBasedInterpolator struct {
 	frozen           bool
-	elpTime, totTime float64
+	elpTime, totTime float32
 	alive            bool
-	val              float64
+	val              float32
 }
 
 func (t *timeBasedInterpolator) Kill()             { t.alive = false }
 func (t *timeBasedInterpolator) IsAlive() bool     { return t.alive }
 func (t *timeBasedInterpolator) Freeze()           { t.frozen = true }
 func (t *timeBasedInterpolator) Thaw()             { t.frozen = false }
-func (t *timeBasedInterpolator) GetValue() float64 { return t.val }
+func (t *timeBasedInterpolator) GetValue() float32 { return t.val }
 
 type LinearTimeInterpolator struct {
 	timeBasedInterpolator
-	startVal, endVal float64
+	startVal, endVal float32
 }
 
 // LinearTimeInterpolator
 //
 // It linearly goes from the start value to the end value over timespan milliseconds
-func NewLinearTimeInterpolator(timespan, start, end float64) *LinearTimeInterpolator {
+func NewLinearTimeInterpolator(timespan, start, end float32) *LinearTimeInterpolator {
 	return &LinearTimeInterpolator{timeBasedInterpolator{false, 0, timespan, true, 0}, start, end}
 }
 
-func clamp(min, max, val float64) float64 {
+func clamp(min, max, val float32) float32 {
 	if val < min {
 		return min
 	} else if val > max {
@@ -69,7 +69,7 @@ func clamp(min, max, val float64) float64 {
 	}
 }
 
-func (l *LinearTimeInterpolator) Update(dT float64) {
+func (l *LinearTimeInterpolator) Update(dT float32) {
 	l.elpTime += dT
 
 	b := clamp(0, 1, l.elpTime/l.totTime)
@@ -82,14 +82,14 @@ func (l *LinearTimeInterpolator) Update(dT float64) {
 
 type QuadraticTimeInterpolator struct {
 	timeBasedInterpolator
-	startVal, endVal, midVal float64
+	startVal, endVal, midVal float32
 }
 
-func NewQuadraticTimeInterpolator(timespan, start, end, mid float64) *QuadraticTimeInterpolator {
+func NewQuadraticTimeInterpolator(timespan, start, end, mid float32) *QuadraticTimeInterpolator {
 	return &QuadraticTimeInterpolator{timeBasedInterpolator{false, 0, timespan, true, 0}, start, end, mid}
 }
 
-func (q *QuadraticTimeInterpolator) Update(dT float64) {
+func (q *QuadraticTimeInterpolator) Update(dT float32) {
 	q.elpTime += dT
 
 	b := clamp(0, 1, q.elpTime/q.totTime)
@@ -103,14 +103,14 @@ func (q *QuadraticTimeInterpolator) Update(dT float64) {
 
 type CubicTimeInterpolator struct {
 	timeBasedInterpolator
-	startVal, endVal, midVal1, midVal2 float64
+	startVal, endVal, midVal1, midVal2 float32
 }
 
-func NewCubicTimeInterpolator(timespan, start, end, mid1, mid2 float64) *CubicTimeInterpolator {
+func NewCubicTimeInterpolator(timespan, start, end, mid1, mid2 float32) *CubicTimeInterpolator {
 	return &CubicTimeInterpolator{timeBasedInterpolator{false, 0, timespan, true, 0}, start, end, mid1, mid2}
 }
 
-func (c *CubicTimeInterpolator) Update(dT float64) {
+func (c *CubicTimeInterpolator) Update(dT float32) {
 	c.elpTime += dT
 
 	b := clamp(0, 1, c.elpTime/c.totTime)
